refactor to hooks
